servers: added nettest interface to DIR service, nettestClient
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@1559 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ src/servers/src/org/xtreemfs/interfaces/NettestInterface/NettestInterface.java
@@ -12,7 +12,7 @@ import yidl.runtime.PrettyPrinter;
 
 public class NettestInterface
 {
-    public static int getProg() { return 2546883425; }
+    public static long getProg() { return 2546883425l; }
     public static int getVersion() { return 2010012513; }
 
     public static Request createRequest( ONCRPCRequestHeader header ) throws Exception
